Game Tic-Tac-Toe [X] [O] [X] [O]
Tic-Tac-Toe is game where two players, X and O, who take turns marking the spaces in a 3×3 grid. The game is designed in such a way that the player enters a number between [1-9] and the game board will be displayed on screen. At the end of the game system will display the winning player.

Game rules:
1. Player 1 plays for [X] --- Player 2 plays for [O]
2. Enter the number between [1-9] to mark the position
3. If all the X's are marked together in a row (up, down, across, or diagonally), then X is the winner.
4. If all the O's are marked together in a row (up, down, across, or diagonally), then O is the winner.
5. When all 9 squares are full, the game is over. If no player has 3 marks in a row, the game ends in a tie.

Conditions:
1. You cannot enter the number apart from [1-9] to mark the position. System displays "Please enter the position between [1-9]" message.
2. You cannot fill the position which has been already filled. System displays "Position has been marked already" message.